Close #405 - replace "temporal" with "longitudinal"
In resolving the change I have put in a written reference to 2 large prospective cohort studies, one US and one UK, as brief examples. The websites of both studies contain bibliographies of results published from the study. I have additionally put in a reference to the seminal Kaplan-Meier paper into the sentence on the importance of the sequence of events; as that paper has been a driving force, for 50 years, in how longitudinal and survival data has been analyzed

I apologize for the line lengths, I edited this online and there is no right-hand margin indicator, so I had to eyeball it.