Update 05.treat.md
New references: 
[@doi:10.1038/s41587-020-0418-2]
[@arxiv:1802.04364]
[@arXiv:1705.10843]
[@arXiv:1806.02473]
[@10.1021/acsmedchemlett.0c00088]
[@doi:10.1021/acs.jcim.0c00174]

note on references - I could not get DOIs for these so had to go with arXiv: 
[@arxiv:1802.04364] is published in ICML , see https://dblp.org/rec/bibtex/conf/icml/JinBJ18  
[@arXiv:1806.02473] is published in NeurIPS, see https://dblp.uni-trier.de/rec/bibtex/conf/nips/YouLYPL18 

Note: another work which led to a synthesized and tested drug molecule is this, from 2018:  https://doi.org/10.1021/acs.molpharmaceut.8b00839. 
However, the 2019 work ( Zhavoronkov et al) we discuss got much more attention. The review is already getting a bit "in the weeds"  so I left it out. 

this is a recent review (july this year) which people may be interested in. I cited it. 
https://pubs.acs.org/doi/pdf/10.1021/acsmedchemlett.0c00088

sorry for any typos that crept in - spellchecker isn't working in Github for some reason.